Improve the circuit latex (source) drawer testing (#5062)
* split test_visualization.py into test_circuit_latex.py and test_utils.py

* reference

* avoid code duplication

* import order
