+++ test/python/visualization/references/test_deep.tex
@@ -1,0 +1,23 @@
+% \documentclass[preview]{standalone}
+% If the image is too large to fit on this documentclass use
+\documentclass[draft]{beamer}
+% img_width = 1, img_depth = 102
+\usepackage[size=custom,height=10,width=159,scale=0.7]{beamerposter}
+% instead and customize the height and width (in cm) to fit.
+% Large images may run out of memory quickly.
+% To fix this use the LuaLaTeX compiler, which dynamically
+% allocates memory.
+\usepackage[braket, qm]{qcircuit}
+\usepackage{amsmath}
+\pdfmapfile{+sansmathaccent.map}
+% \usepackage[landscape]{geometry}
+% Comment out the above line if using the beamer documentclass.
+\begin{document}
+
+\begin{equation*}
+    \Qcircuit @C=1.0em @R=0.0em @!R {
+	 	\lstick{ {q}_{0} :  } & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \gate{H} & \qw & \qw\\
+	 }
+\end{equation*}
+
+\end{document}
