+++ test/python/visualization/references/test_tiny.tex
@@ -1,7 +1,7 @@
 % \documentclass[preview]{standalone}
 % If the image is too large to fit on this documentclass use
 \documentclass[draft]{beamer}
-% img_width = 3, img_depth = 3
+% img_width = 1, img_depth = 3
 \usepackage[size=custom,height=10,width=10,scale=0.7]{beamerposter}
 % instead and customize the height and width (in cm) to fit.
 % Large images may run out of memory quickly.
@@ -14,12 +14,9 @@
 % Comment out the above line if using the beamer documentclass.
 \begin{document}
 
-{\small Global Phase: $\frac{\pi}{2}$}
 \begin{equation*}
     \Qcircuit @C=1.0em @R=0.0em @!R {
 	 	\lstick{ {q}_{0} :  } & \gate{H} & \qw & \qw\\
-	 	\lstick{ {q}_{1} :  } & \gate{H} & \qw & \qw\\
-	 	\lstick{ {q}_{2} :  } & \gate{H} & \qw & \qw\\
 	 }
 \end{equation*}
